backend/requirements.txt
# FastAPI backend
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0
databases>=0.9.0

# Claude Agent SDK (already in .venv)
claude-agent-sdk>=0.1.0

# Utilities
pydantic>=2.5.0
httpx>=0.26.0
